telegram-bot/requirements.txt
python-telegram-bot==20.3
httpx==0.27.0
pydantic==1.10.13